vi_e2e/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==5.0.1
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==2.1.1
click==8.1.3
conda==4.12.0
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
einops==0.4.1
entrypoints==0.4
executing==1.0.0
faiss-cpu==1.7.2
faiss-gpu==1.7.2
fastjsonschema==2.16.1
filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata==4.12.0
importlib-resources==5.9.0
install==1.3.5
ipykernel==6.15.2
ipython==8.5.0
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2==3.0.0
jsonschema==4.15.0
jupyter-client==7.3.5
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict==6.0.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603570489231/work
oauthlib==3.2.0
olefile==0.46
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.7.0
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
pimm==0.0.5
ping3==4.0.3
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1602688806899/work
protobuf==3.19.4
psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyDeprecate==0.3.1
Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-etcd==0.4.5
pytorch-lightning==1.5.7
pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyYAML==5.3.1
pyzmq==23.2.1
requests==2.25.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml==0.15.87
scipy==1.9.0
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.9.8
setproctitle==1.3.2
shortuuid==1.0.9
six @ file:///tmp/build/80754af9/six_1605205327372/work
smmap==5.0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-logger==0.1.0
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
timm==0.5.4
tinycss2==1.1.1
torch==1.7.1
torchelastic==0.2.1
torchmetrics==0.9.3
torchvision==0.8.2
tornado==6.2
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.3.0
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.11
wandb==0.13.2
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1
